chore: some ui structure, node/electron update
