+++ .github/workflows/build.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Install Node.js, NPM and Yarn
         uses: actions/setup-node@v2
         with:
-          node-version: 16.5
+          node-version: 16.13
 
       - name: Build/release Electron app
         uses: samuelmeuli/action-electron-builder@v1
