+++ .github/workflows/ci.yml
@@ -13,7 +13,7 @@ jobs:
       fail-fast: true
       matrix:
         node-version:
-          - 16.5
+          - 16.13
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v2
