+++ src/ui2/contexts.js
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+export const ColorModeContext = createContext({
+  toggleColorMode() {},
+  setColorMode() {},
+});
