+++ src/ui2/main-content.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Box from '@mui/material/Box';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
+
+const MainContent = () => (
+  <Box>
+    <Card>
+      <CardContent>
+        MAIN CONTENT
+      </CardContent>
+    </Card>
+  </Box>
+);
+
+export default MainContent;
