+++ src/ui2/settings/settings.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Box } from '@mui/material';
+
+const Settings = () => (
+  <Box>
+    Settings
+  </Box>
+);
+
+export default Settings;
